refactor bugfix 修改Gain由多声道增益改为全声道增益 Config类静态化，减少内存占用，避免多个实例造成的错误传参 MusicItemModel大量修改，减少内存占用，优化字段数量 MusicListModel重命名为PlaylistModel，减少内存占用，优化内容 使用数据库处理音频文件，处理速度更快，储存更轻便 经过此次修改，Config实现了可向后兼容性和可自定义性，使数据移动后仍能以简单的方式恢复先前的设置

+++ QwQ Music/Services/Audio/AudioPlay.cs
@@ -180,13 +180,13 @@ public class AudioPlay
         PositionChanged?.Invoke(this, _audioFileReader.CurrentTime.TotalSeconds);
     }
 
-    public void PlayWithFade(int fadeInDuration)
+    public void PlayWithFade(ulong fadeInDuration)
     {
         _fadeEffect?.BeginFadeIn(fadeInDuration);
         Play();
     }
 
-    public void StopWithFade(int fadeOutDuration)
+    public void StopWithFade(ulong fadeOutDuration)
     {
         if (_fadeEffect == null)
         {
